% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructions.R
\docType{data}
\name{audio}
\alias{audio}
\title{Instructions & AWS S3 directory
This directory should be changed to where it should be hosted}
\format{
An object of class \code{character} of length 1.
}
\usage{
audio
}
\description{
Instructions & AWS S3 directory
This directory should be changed to where it should be hosted
}
\keyword{datasets}
