% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_buttons.R
\docType{data}
\name{button_style}
\alias{button_style}
\title{This is the CSS styler for the clickable buttons}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
button_style
}
\description{
This is the CSS styler for the clickable buttons
}
\keyword{datasets}
