% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_blocks.R
\name{envBlock}
\alias{envBlock}
\title{all three testing blocks that join demo and real sessions}
\usage{
envBlock(with_training = TRUE, dict = tptR::TPT_dict)
}
\description{
all three testing blocks that join demo and real sessions
}
