% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructions.R
\name{instructions}
\alias{instructions}
\title{Instructions & AWS S3 directory}
\usage{
instructions(with_volume_headphone_check = FALSE, dict = tptR::TPT_dict)
}
\arguments{
\item{with_volume_headphone_check}{page asking to calibrate volume and
a question about audio output device. Default is \code{FALSE}}

\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
Instructions & AWS S3 directory
}
