% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_demo_page.R
\name{makeDemoPage}
\alias{makeDemoPage}
\title{Makes demo practice trials}
\usage{
makeDemoPage(blockName, trialName, targetValue, dict = tptR::TPT_dict)
}
\arguments{
\item{blockName}{name of the block to be displayed e.g. Block1}

\item{trialName}{name of audio file prefix. If file is called 'env_13.mp3', 'env' is the trialName}

\item{targetValue}{the value of sampled slider position. If file is called 'env_13.mp3', '13' is the targetValue}

\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
This assigns where the target value (i.e. correct slider position) should be in a given
practice trial.
}
