% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_test_page.R
\name{makeTestPage}
\alias{makeTestPage}
\title{Make real testing trials}
\usage{
makeTestPage(trialName, blockName)
}
\arguments{
\item{trialName}{trial items to be sourced. Eg, env1, env2, ... env6}

\item{blockName}{block name to be displayed to participant. Eg, Block1}
}
\description{
Make real testing trials
}
