% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_blocks.R
\docType{data}
\name{page_after_block}
\alias{page_after_block}
\title{all three testing blocks that join demo and real sessions}
\format{
An object of class \code{list} of length 3.
}
\usage{
page_after_block
}
\description{
all three testing blocks that join demo and real sessions
}
\keyword{datasets}
