% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_slider.R
\name{sliderJS}
\alias{sliderJS}
\title{This custom script for moveable slider that records slider position}
\usage{
sliderJS(trialName)
}
\arguments{
\item{trialName}{If file is called 'env_13.mp3', 'env' is the trialName and the function maps
slider values from 0 to 100. Eg, env_01, env_02, env_03...}
}
\description{
This custom script for moveable slider that records slider position
}
