% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPT.R
\name{TPT}
\alias{TPT}
\title{#' TPT module}
\usage{
TPT(
  label = "TPT",
  with_welcome = TRUE,
  with_training = TRUE,
  with_feedback = TRUE,
  dict = tptR::TPT_dict,
  ...
)
}
\arguments{
\item{label}{(Scalar character) Label for saving data}

\item{with_welcome}{(Scalar boolean) Indicates if welcome page will be shown  at beginning of the test module.
Defaults to TRUE}

\item{with_training}{(Boolean scalar) Defines whether instructions and training are included.
Defaults to TRUE.}

\item{with_feedback}{(Scalar boolean) Indicates if performance feedback will be given at the end of the test.
Defaults to TRUE}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{...}{Further arguments to be passed to \code{\link{TPT}()}.}
}
\description{
This function contains a TPT psychTestR-module for inlcusion into batter
}
