% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpt_demo.R
\name{TPT_demo}
\alias{TPT_demo}
\title{TPT_demo}
\usage{
TPT_demo(password = "demo", your_email = "your-email@address")
}
\arguments{
\item{password}{default as 'demo'}

\item{your_email}{default as 'your-email@address'}
}
\description{
This is a quick demo version of the TPT. Only a single testing block is displayed
with a fake feedback score of 80 at the end.
}
