% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpt_standalone.R
\name{TPT_standalone}
\alias{TPT_standalone}
\title{TPT_standalone}
\usage{
TPT_standalone(password, your_email = "hlee@cbs.mpg.de")
}
\arguments{
\item{password}{Create a password for logging in as an admin.}

\item{your_email}{The researcher's email to be displayed. Replace with your email address.}
}
\description{
Run TPT as a standalone (i.e. run locally without hosting on a server).
Running this locally will create a local directory folder called 'output' and
store participant's response there.
}
