% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpt_standalone.R
\name{TPT_test}
\alias{TPT_test}
\title{TPT_test}
\usage{
TPT_test(password, researcher_email = "hlee@cbs.mpg.de")
}
\arguments{
\item{password}{Create a password to login as an admin}

\item{researcher_email}{The researcher's email to be displayed. Default set to the
developer's email}
}
\description{
Run TPT as a standalone (i.e. run locally without hosting on a server).
Running this locally will create a local directory called 'output' and
store participant's response there.
}
\examples{
TPT_test(password = "test1234", researcher_email = "hlee@cbs.mpg.de")
}
