% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{TPT_demo}
\alias{TPT_demo}
\title{Demo TPT}
\usage{
TPT_demo(
  take_training = TRUE,
  admin_password = "demo",
  researcher_email = "longgold@gold.uc.ak",
  dict = tptR::TPT_dict,
  language = "en",
  feedback = TRUE
)
}
\arguments{
\item{take_training}{(Boolean scalar) Defines whether instructions and training are included.
Defaults to TRUE.}

\item{admin_password}{(Scalar character) Password for accessing the admin panel.
Defaults to \code{"demo"}.}

\item{researcher_email}{(Scalar character)
If not \code{NULL}, this researcher's email address is displayed
at the bottom of the screen so that online participants can ask for help.
Defaults to \email{longgold@gold.uc.ak},
the email address of this package's developer.}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{language}{The language you want to run your demo in.
Possible languages include English (\code{"en"}) and German (informal \code{"de"}, formal \code{"de_f"}).
The first language is selected by default}

\item{feedback}{(Function) Defines the feedback to give the participant
at the end of the test. Defaults to a graph-based feedback page.}

\item{...}{Further arguments to be passed to \code{\link{TPT}()}.}
}
\description{
This function launches a demo for the TPT
}
