% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpt_standalone.R
\name{TPT_standalone}
\alias{TPT_standalone}
\title{TPT_standalone}
\usage{
TPT_standalone(password, researcher_email = "hlee@cbs.mpg.de")
}
\arguments{
\item{password}{Create a password to login as an admin}

\item{researcher_email}{The researcher's email to be displayed. Default set to the
developer's email}
}
\description{
Run TPT as a standalone (i.e. run locally without hosting on a server).
Running this locally will create a local directory folder called 'output' and
store participant's response there.
}
